+++ source/pythonpath/easydev/comun.py
@@ -4,6 +4,8 @@ import logging
 import datetime
 import os
 import tempfile
+import zipfile
+from functools import wraps
 
 import uno
 from com.sun.star.beans import PropertyValue, NamedValue
@@ -24,6 +26,16 @@ if PY2:
 log = logging.getLogger(NAME_EXT)
 CTX = uno.getComponentContext()
 SM = CTX.getServiceManager()
+
+
+def catch_exception(f):
+    @wraps(f)
+    def func(*args, **kwargs):
+        try:
+            return f(*args, **kwargs)
+        except Exception as e:
+            log.error(f.__name__, exc_info=True)
+    return func
 
 
 def _create_instance(name, with_context=True):
@@ -66,7 +78,7 @@ def replace_filename(path, filename):
 
 def replace_name_ext(path_target, path_source, ext):
     _, _, name, _ = get_path_info(path_source)
-    return os.path.join(path_target, '{}.{}'.format(path, name, ext))
+    return os.path.join(path_target, '{}.{}'.format(name, ext))
 
 
 def get_path_info(path):
@@ -264,6 +276,58 @@ def get_pos_size(data):
     pos.X = data.X
     pos.Y = data.Y
     return pos, size
+
+
+def set_property(control, properties):
+    uno.invoke(control, 'setPropertyValue', properties)
+    return
+
+
+def zip_file(source, target):
+    if not target:
+        target = replace_ext(source, 'zip')
+    elif isdir(target):
+        target = replace_name_ext(target, source, 'zip')
+    z = zipfile.ZipFile(target, 'w', compression=zipfile.ZIP_DEFLATED)
+    if exists(source):
+        _, name, _, _ = get_path_info(source)
+        z.write(source, name)
+    z.close()
+    return
+
+
+def zip_dir(source, target):
+    if not target:
+        target = path_to_url(source).split('/')
+        target[-1] = '{}.zip'.format(target[-1])
+        target = path_to_os('/'.join(target))
+    elif isdir(target):
+        path = path_to_url(source).split('/')
+        target = join(target, '{}.zip'.format(path[-1]))
+    z = zipfile.ZipFile(target, 'w', compression=zipfile.ZIP_DEFLATED)
+    root_len = len(os.path.abspath(source))
+    for root, dirs, files in os.walk(source):
+        relative = os.path.abspath(root)[root_len:]
+        for f in files:
+            fullpath = join(root, f)
+            file_name = join(relative, f)
+            z.write(fullpath, file_name)
+    z.close()
+    return
+
+
+def unzip(source, target, name):
+    z = zipfile.ZipFile(source)
+    if name:
+        z.extract(name, target)
+    else:
+        if not target:
+            path, _, name, _ = get_path_info(source)
+            target = join(path, name)
+        z.extractall(target)
+    z.close()
+    return
+
     #~ new_shape = doc.createInstance(SRV_GOS)
     #~ dp.add(new_shape)
     #~ new_shape.Graphic = src_img.Graphic